+++ packages/cli/src/info/env_info_provider.ts
@@ -1,4 +1,4 @@
-import * as os from 'node:os';
+import { EOL } from 'os';
 import envinfo from 'envinfo';
 import { format } from '@aws-amplify/cli-core';
 import { EnvInfo } from './env_info_provider_types.js';
@@ -16,15 +16,33 @@ export class EnvironmentInfoProvider {
         System: ['OS', 'CPU', 'Memory', 'Shell'],
         Binaries: ['Node', 'Yarn', 'npm', 'pnpm'],
         npmPackages: [
+          '@aws-amplify/auth-construct',
           '@aws-amplify/backend',
+          '@aws-amplify/backend-auth',
           '@aws-amplify/backend-cli',
+          '@aws-amplify/backend-data',
+          '@aws-amplify/backend-deployer',
+          '@aws-amplify/backend-function',
+          '@aws-amplify/backend-output-schemas',
+          '@aws-amplify/backend-output-storage',
+          '@aws-amplify/backend-secret',
+          '@aws-amplify/backend-storage',
+          '@aws-amplify/cli-core',
+          '@aws-amplify/client-config',
+          '@aws-amplify/deployed-backend-client',
+          '@aws-amplify/form-generator',
+          '@aws-amplify/model-generator',
+          '@aws-amplify/platform-core',
+          '@aws-amplify/plugin-types',
+          '@aws-amplify/sandbox',
+          '@aws-amplify/schema-generator',
           'typescript',
           'aws-cdk',
           'aws-cdk-lib',
           'aws-amplify',
         ],
       },
-      { json: true, showNotFound: true }
+      { json: true, showNotFound: true, fullTree: true }
     );
     return this.formatEnvInfo(JSON.parse(info));
   }
@@ -53,14 +71,25 @@ export class EnvironmentInfoProvider {
     const npmPackages = [
       'NPM Packages:',
       ...Object.entries(info.npmPackages).map(([name, details]) => {
-        if (typeof details !== 'string') {
-          return format.indent(`${name}: ${details.installed}`);
+        // envinfo with fullTree: true returns the version of the project package.json under %name% which we don't care about
+        if (name === '%name%') {
+          return;
         }
-        return format.indent(`${name}: ${details}`);
+        if (typeof details === 'string') {
+          return format.indent(`${name}: ${details}`);
+        } else if (Object.keys(details).length === 0) {
+          // if package details are empty, exclude it from the output. Note that this is different than the package not being installed
+          // the potential for empty details seems to be a side-effect of fullTree: true on envinfo
+          return;
+        }
+        return format.indent(`${name}: ${details.installed}`);
       }),
     ];
 
-    const output = [system, binaries, npmPackages].flat().join(os.EOL);
+    const output = [system, binaries, npmPackages]
+      .flat()
+      .filter((line) => !!line)
+      .join(EOL);
     return output;
   }
 }